[TASK] fine tuning of Configuration task
The Configuration task now is seamlessly integrated into the Deploy
task. The LocalConfiguration.properties exist in as a template and
get extended by the Configuration task. Furthermore the Configuration
task asks if the user wants to skip deployment in favor of taking
manual adjustments of the LocalConfiguration.properties file if the
file was generated the first time.